% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallUtils.R
\name{as.data.frame.coxsim}
\alias{as.data.frame.coxsim}
\title{Convert a coxsim class object into a data frame}
\usage{
\method{as.data.frame}{coxsim}(x, ...)
}
\arguments{
\item{x}{a \code{coxsim} class object.}

\item{...}{arguments to pass to \code{data.frame}.}
}
\description{
Convert a coxsim class object into a data frame
}
