% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{GolubEUPData}
\alias{GolubEUPData}
\title{A data set from Golub & Steunenberg (2007)}
\format{A data set with 3001 observations and 17 variables.}
\source{
Golub, Jonathan, and Bernard Steunenberg. 2007. ''How Time Affects
EU Decision-Making.'' European Union Politics 8(4): 555-66.

Amanda A. Licht, 2011, "Replication data for: Change Comes with Time".
\url{http://hdl.handle.net/1902.1/15633}. IQSS Dataverse Network
[Distributor] V3 [Version].
}
\description{
A data set from Golub & Steunenberg (2007)
}
\keyword{datasets}

