% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmohivData.R
\docType{data}
\name{hmohiv}
\alias{hmohiv}
\title{Simulated HIV patient data from UCLA IDRE}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\source{
UCLA IDRE \url{https://stats.oarc.ucla.edu/r/examples/asa/r-applied-survival-analysis-ch-1/}
}
\usage{
hmohiv
}
\description{
Simulated HIV patient data from UCLA IDRE
}
\keyword{datasets}
