% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/smallUtils.R
\name{setXl}
\alias{setXl}
\title{Create a sequence of Xl values}
\usage{
setXl(Xj, diff = 1)
}
\arguments{
\item{Xj}{numeric vector of fitted values for the covariate of interest to
simulate for.}

\item{diff}{numeric vector of length 1. It specifies the difference between
\code{Xj} and \code{Xl}. \code{Xl} is always smaller than \code{Xj}.}
}
\value{
a vector
}
\description{
\code{setXl} creates a sequence of \code{Xl} values given a sequence of
\code{Xj} values and a fixed difference.
}
\examples{
# Set Xj
setXj = seq(1100, 1700, by = 10)

# Find Xl that are 1 less than Xj
setXl(Xj = setXj, diff = 1)
}
\keyword{utilities}

