\name{SpinMakeH}
\alias{SpinMakeH}
\title{Internal function for finding the upper bound from SPIn}
\usage{
  SpinMakeH(y)
}
\arguments{
  \item{y}{the name of the simulated quantities of interest
  variable}

  \item{ci}{numeric confidence level}
}
\description{
  \code{SpinMakeL} internal function for finding the lower
  bound from SPIn
}

