\name{SpinMakeL}
\alias{SpinMakeL}
\title{Internal function for finding the lower bound from SPIn}
\usage{
  SpinMakeL(y, ci)
}
\arguments{
  \item{y}{the name of the simulated quantities of interest
  variable}

  \item{ci}{numeric confidence level}
}
\description{
  \code{SpinMakeL} internal function for finding the lower
  bound from SPIn
}

