% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MinMaxLines}
\alias{MinMaxLines}
\title{Transform the simulation object to include only the min and max of the
constricted interval, as well as the lower and upper bounds of the middle 50
percent of the constricted interval}
\usage{
MinMaxLines(df, byVars = "Xj", hr = FALSE, strata = FALSE,
  clean = FALSE)
}
\arguments{
\item{df}{a data frame or a simulation class object.}

\item{byVars}{character vector of the variables to subset the data frame by.
The default is \code{'Xj'}.}

\item{hr}{logical indicating whether or not \code{df} contains a hazard rate.}

\item{strata}{logical indicating whether or not \code{df} contains a
stratified hazard rate.}

\item{clean}{logical, whether or not to clean up the output data frame to
only include \code{byVars}, \code{Min_CI}, \code{Lower50_CI}, \code{median},
\code{Upper50_CI}, \code{Max_CI}.}
}
\description{
\code{MinMaxLines} is an internal function to transform the simulation
object to include only the min and max of the interval set by \code{ci} in
the \code{coxsim} command, as well as the lower and upper bounds of the
middle 50 percent of this interval. It also returns the median.
}
\keyword{internals}

