\name{simGG}
\alias{simGG}
\title{A method for plotting simulation objects created by simPH.}
\usage{
  simGG(obj, ...)
}
\description{
  A method for plotting simulation objects created by
  simPH.
}
\seealso{
  \code{\link{simGG.siminteract}},
  \code{\link{simGG.simtv}}, \code{\link{simGG.simlinear}},
  \code{\link{simGG.simpoly}},
  \code{\link{simGG.simspline}}
}

