\name{simGG}
\alias{simGG}
\title{A method for plotting simulation objects created by simPH.}
\usage{
  simGG(obj, ...)
}
\arguments{
  \item{obj}{an object created by one of simPH's simulation
  commands.}

  \item{...}{arguments to be passed to methods.}
}
\description{
  A method for plotting simulation objects created by
  simPH.
}
\seealso{
  \code{\link{simGG.siminteract}},
  \code{\link{simGG.simtvc}},
  \code{\link{simGG.simlinear}},
  \code{\link{simGG.simpoly}},
  \code{\link{simGG.simspline}}
}

